fix: remove preview and adjustments to the preview component
