--- src/components/ComponentPreview/PreviewContainer.stories.tsx
@@ -1,13 +1,0 @@
-import { Story } from '@ladle/react';
-
-import Preview from '~components/Preview/Preview';
-
-import PreviewContainer from './PreviewContainer';
-
-export const PreviewContainerStories: Story = () => (
-  <div>
-    <PreviewContainer>
-      <Preview>Preview Imaginário</Preview>
-    </PreviewContainer>
-  </div>
-);
