--- src/components/ComponentPreview/PreviewContainer.tsx
@@ -1,11 +1,0 @@
-import { ReactNode } from 'react';
-
-import scss from './PreviewContainer.module.scss';
-
-import { IPreviewContainerProps } from './PreviewContainer.types';
-
-function PreviewContainer(props: IPreviewContainerProps): ReactNode {
-  return <div className={scss.container}>{props.children}</div>;
-}
-
-export default PreviewContainer;
