--- src/components/ComponentPreview/PreviewContainer.types.ts
@@ -1,5 +1,0 @@
-import { ReactNode } from 'react';
-
-export type IPreviewContainerProps = {
-  children: ReactNode;
-};
