+++ src/components/Preview/Preview.module.scss
