+++ src/components/Preview/Preview.spec.tsx
@@ -3,11 +3,16 @@ import { render, screen } from '@testing-library/react';
 import Preview from './Preview';
 
 describe('Preview', () => {
-  describe('when pass the preview props', () => {
-    it('renders on screen the preview', () => {
-      render(<Preview>test</Preview>);
-      const preview = screen.getByText(/test/i);
-      expect(preview).toBeInTheDocument();
+  describe('when initialized', () => {
+    it('renders the element preview inside the container', () => {
+      render(
+        <Preview>
+          <h1> teste </h1>
+        </Preview>
+      );
+
+      const titleComponent = screen.getByText('teste');
+      expect(titleComponent).toBeInTheDocument();
     });
   });
 });
