+++ src/components/Preview/Preview.types.ts
@@ -1,5 +1,5 @@
 import { ReactNode } from 'react';
 
-export type IPreview = {
+export type IPreviewContainerProps = {
   children: ReactNode;
 };
